+++ examples/data-sources/sdwan_policy_object_security_scalable_group_tag_profile_parcel/data-source.tf
@@ -1,0 +1,4 @@
+data "sdwan_policy_object_security_scalable_group_tag_profile_parcel" "example" {
+  id                 = "f6b2c44c-693c-4763-b010-895aa3d236bd"
+  feature_profile_id = ""
+}