+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - Add `sdwan_policy_object_security_ips_signature_profile_parcel` resource and data source
 - Add `sdwan_policy_object_security_localdomain_profile_parcel` resource and data source
 - Add `sdwan_policy_object_security_port_profile_parcel` resource and data source
+- Add `sdwan_policy_object_security_scalable_group_tag_profile_parcel` resource and data source
 
 ## 0.3.13